Updated for more universal support

+++ bootstrap.sh
@@ -1,9 +1,13 @@
 #!/usr/bin/env bash
 set -e
 
+# Installation Variables
+VENV=metrics-env
+
+
 sudo apt-get update
-sudo apt-get install -y python3 python3-venv python3-pip
-sudo apt-get install -y virtualenv
+sudo apt-get upgrade -y
+sudo apt-get install -y python3 virtualenv python3-pip
 
 # Prepare Django Environment
 cd /vagrant
@@ -14,15 +18,29 @@ if [ ! -d cscimetrics ]; then
 fi
 cd cscimetrics
 
-if [ ! -d metrics-env ]; then
-    echo -  Creating new virtual environment - metrics-env
-    virtualenv -p /usr/bin/python3 metrics-env --always-copy
+# Configure Python Virtual Environment
+if [ ! -d $VENV ]; then
+    echo - Creating new virtual environment - $VENV
+    virtualenv -p $(which python3) $VENV --always-copy
+
+else
+    echo - Cleaning old virtual environment - $VENV
+    rm -rf $VENV
+    virtualenv -p $(which python3) $VENV --always-copy
 fi
 
+# Install Python Package Dependencites
+echo - Installing Python Packages
+source $VENV/bin/activate
+
+if [ -e requirements.txt ]; then
+    pip install -r requirements.txt
+else
+    echo - Missing \"requirements.txt\". Please install requirements manually and execute \"pip freeze > requirements.txt\"
+fi
+
+# Configure SSH Login Parameters
 echo ". /vagrant/cscimetrics/metrics-env/bin/activate" > ~/.profile
 echo "cd /vagrant/cscimetrics" >> ~/.profile
-
-source metrics-env/bin/activate
-pip install -r requirements.txt
 
 echo Setup Complete!
